Add links to places where we can report issues
This PR adds an incomplete section to the README file with links to an issue repository where issues with PFB and pypfb can be reported. I would have reported this as a GitHub issue, but your GitHub issues are not accessible to non-project users, so hopefully someone can fill in the correct URLs to use in the README so that we'll know where to report issues in the future :).